modules: Move some types out of py_image.
